Create Q3
This is a project created by Thomas Herndon, Michael Ash, and Robert Pollin analyzing the growth of GDP of multiple nations in debt. The authors compared the GDP growth with respect to the debt percentage according to the nation's GDP. The interesting thing about the study is that the percent of debt seems not to have a strong correlation with GDP growth. Another thing that amazed me is how strong and efficient machine and coding can be while dealing with and arranging large data sets. This makes it possible for economists to use actual data to forecast economies and find relationships between different economic phenomenons.